+++ app/code/Magento/Payment/Model/Cart.php
@@ -16,9 +16,13 @@ class Cart
      * Amounts
      */
     const AMOUNT_TAX = 'tax';
+
     const AMOUNT_SHIPPING = 'shipping';
+
     const AMOUNT_DISCOUNT = 'discount';
+
     const AMOUNT_SUBTOTAL = 'subtotal';
+
     /**@@+*/
 
     /**
@@ -307,8 +311,11 @@ class Cart
                 continue;
             }
 
-            $this->_salesModelItems[] = $this->_createItemFromData($item->getName(), $item->getQty(),
-                $item->getPrice());
+            $this->_salesModelItems[] = $this->_createItemFromData(
+                $item->getName(),
+                $item->getQty(),
+                $item->getPrice()
+            );
         }
 
         $this->addSubtotal($this->_salesModel->getBaseSubtotal());
@@ -357,7 +364,7 @@ class Cart
      */
     protected function _setAmount($amountType, $amount)
     {
-        $this->_amounts[$amountType] = (float)$amount;
+        $this->_amounts[$amountType] = (double)$amount;
     }
 
     /**
@@ -367,7 +374,7 @@ class Cart
      */
     protected function _addAmount($amountType, $amount)
     {
-        $this->_amounts[$amountType] += (float)$amount;
+        $this->_amounts[$amountType] += (double)$amount;
     }
 
     /**
@@ -392,11 +399,7 @@ class Cart
      */
     protected function _createItemFromData($name, $qty, $amount, $identifier = null)
     {
-        $item = new \Magento\Object(array(
-            'name'   => $name,
-            'qty'    => $qty,
-            'amount' => (float)$amount,
-        ));
+        $item = new \Magento\Object(array('name' => $name, 'qty' => $qty, 'amount' => (double)$amount));
 
         if ($identifier) {
             $item->setData('id', $identifier);
@@ -416,7 +419,7 @@ class Cart
             self::AMOUNT_DISCOUNT => 0,
             self::AMOUNT_SHIPPING => 0,
             self::AMOUNT_SUBTOTAL => 0,
-            self::AMOUNT_TAX      => 0
+            self::AMOUNT_TAX => 0
         );
     }
 }
